декомпозиция проекта
