+++ 3-struct/file/file.go
@@ -1,0 +1,1 @@
+package file
